+++ packages/react-magma-docs/src/components/layout.js
@@ -8,6 +8,7 @@ import './layout.css';
 import LayoutComponent from './layout-component';
 import editorTheme from './editorTheme';
 import { v4 as uuid } from 'uuid';
+import { SimplePropsTable } from './props-table';
 
 const PreComponent = ({ className, components, ...props }) => {
   const hideCode = props.children.props.hideCode;
@@ -74,6 +75,7 @@ const Layout = ({ children, pageContext }) => {
     pageContext && pageContext.frontmatter
       ? pageContext.frontmatter.pageTitle || pageContext.frontmatter.title || ''
       : '';
+
   return (
     <LayoutComponent title={title}>
       <MDXProvider
@@ -92,27 +94,28 @@ const Layout = ({ children, pageContext }) => {
 };
 
 export const ScopeableLayout = ({ children, components, pageContext }) => {
-  const title =
-    pageContext && pageContext.frontmatter
-      ? pageContext.frontmatter.pageTitle || pageContext.frontmatter.title || ''
-      : '';
+  const properties = (pageContext && pageContext.properties) || [];
+
   return (
-    <LayoutComponent title={title}>
-      <MDXProvider
-        components={{
-          pre: preProps => (
-            <PreComponent {...preProps} components={components} />
-          ),
-          table: Table,
-          h2: SectionHeading,
-          h3: LinkHeading,
-        }}
-      >
-        <article className="content-article">
-          <SkipLinkContent>{children}</SkipLinkContent>
-        </article>
-      </MDXProvider>
-    </LayoutComponent>
+    <MDXProvider
+      components={{
+        pre: preProps => <PreComponent {...preProps} components={components} />,
+        table: Table,
+        h2: SectionHeading,
+        h3: LinkHeading,
+        SimplePropsTable: SimplePropsTable,
+        ...properties.reduce((acc, { name, properties }) => {
+          return {
+            ...acc,
+            [name]: args => (
+              <SimplePropsTable propertyValues={properties} {...args} />
+            ),
+          };
+        }, {}),
+      }}
+    >
+      {children}
+    </MDXProvider>
   );
 };