+++ packages/react-magma-dom/src/components/Banner/index.tsx
@@ -9,9 +9,23 @@ import { IconButton } from '../IconButton';
 import { ThemeContext } from '../../theme/ThemeContext';
 import { I18nContext } from '../../i18n';
 
+/**
+ * @children required
+ */
 export interface BannerProps extends AlertProps {
+  /**
+   * The text displayed inside of the action button
+   */
   actionButtonText?: string;
+  /**
+   * Action that fires when the action button is clicked. Must be present for button to appear
+   */
   actionButtonOnClick?: () => void;
+  /**
+   * If true, the component will be able to be dismissed and will include a close button
+   * @default false
+   */
+  isDismissible?: boolean;
 }
 
 const StyledBanner = styled.div<AlertProps>`